refactor(pb): add version to package
