+++ gen/proto/services/auth_svc/v1/logging.go
