+++ gen/proto/services/emergency_room_svc/v1/helpers.go
