+++ images/service-preproc/run.sh
@@ -8,5 +8,5 @@ shift # move the variable names, $2 -> $1, $3 -> $2, ... "$@" now ignores the ol
 buf $CMD $@
 
 if [ "$CMD" = "generate" ]; then
-	protoc-go-inject-tag -input="gen/proto/*/*/*.pb.go" # Note this only matches exactly two dirs, as go's glob does not support '**' (https://github.com/golang/go/issues/11862)
+	protoc-go-inject-tag -input="gen/proto/*/*/*/*.pb.go" # Note this only matches exactly three dirs, as go's glob does not support '**' (https://github.com/golang/go/issues/11862)
 fi
