--- proto/libs/events/user_events.proto
@@ -1,11 +1,0 @@
-syntax = "proto3";
-
-option go_package = "gen/proto/libs/events";
-
-// TOPIC: USER_CREATED
-message UserCreatedEvent {
-	string id = 1;
-	string email = 2;
-	string nickname = 3;
-	string fullName = 4;
-}
