+++ proto/libs/events/v1/user_events.proto
@@ -1,0 +1,13 @@
+syntax = "proto3";
+
+package proto.libs.events.v1;
+
+option go_package = "gen/proto/libs/events";
+
+// TOPIC: USER_CREATED
+message UserCreatedEvent {
+	string id = 1;
+	string email = 2;
+	string nickname = 3;
+	string fullName = 4;
+}
