--- proto/services/task-svc/room_svc.proto
@@ -1,67 +1,0 @@
-syntax = "proto3";
-
-option go_package = "gen/proto/services/task-svc";
-
-//
-// CreateRoom
-//
-
-message CreateRoomRequest {
-	string name = 1; // @gotags: validate:"required"
-	uint32 amountOfBeds = 2; // @gotags: validate:"required"
-}
-
-message CreateRoomResponse {
-	string id = 1;
-}
-
-//
-// GetRoom
-//
-
-message BedOfRoom {
-	string id = 1;
-}
-
-message GetRoomRequest {
-	string id = 1; // @gotags: validate:"uuid4"
-}
-
-message GetRoomResponse {
-	string id = 1;
-	string name = 2;
-	repeated BedOfRoom beds = 3;
-}
-
-//
-// UpdateRoom
-//
-
-message UpdateRoomRequest {
-	string id = 1; // @gotags: validate:"uuid4"
-	optional string name = 2;
-}
-
-message UpdateRoomResponse {}
-
-//
-// AddBedsToRoom
-//
-
-message AddBedsToRoomRequest {
-	string id = 1; // @gotags: validate:"uuid4"
-	uint32 amountOfBeds = 3;
-}
-
-message AddBedsToRoomResponse {}
-
-//
-// RoomService
-//
-
-service RoomService {
-	rpc CreateRoom(CreateRoomRequest) returns (CreateRoomResponse) {}
-	rpc GetRoom(GetRoomRequest) returns (GetRoomResponse) {}
-	rpc UpdateRoom(UpdateRoomRequest) returns (UpdateRoomResponse) {}
-	rpc AddBedsToRoom(AddBedsToRoomRequest) returns (AddBedsToRoomResponse) {}
-}
