+++ proto/services/task_svc/v1/patient_svc.proto
@@ -1,0 +1,51 @@
+syntax = "proto3";
+
+package proto.services.task_svc.v1;
+
+option go_package = "gen/proto/services/task-svc";
+
+//
+// CreatePatient
+//
+
+message CreatePatientRequest {
+	string humanReadableIdentifier = 1; // @gotags: validate:"required"
+}
+
+message CreatePatientResponse {
+	string id = 1;
+}
+
+//
+// GetPatient
+//
+
+message GetPatientRequest {
+	string id = 1; // @gotags: validate:"uuid4"
+}
+
+message GetPatientResponse {
+	string id = 1; // @gotags: validate:"uuid4"
+	string humanReadableIdentifier = 2; // @gotags: validate:"required"
+}
+
+//
+// UpdatePatient
+//
+
+message UpdatePatientRequest {
+	string id = 1; // @gotags: validate:"uuid4"
+	optional string humanReadableIdentifier = 2;
+}
+
+message UpdatePatientResponse {}
+
+//
+// PatientService
+//
+
+service PatientService {
+	rpc CreatePatient(CreatePatientRequest) returns (CreatePatientResponse) {}
+	rpc GetPatient(GetPatientRequest) returns (GetPatientResponse) {}
+	rpc UpdatePatient(UpdatePatientRequest) returns (UpdatePatientResponse) {}
+}
